+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.0.6
+- Theme aware struct tag colors: struct tags now use the colors from the current theme.
+- New configuration `goStructTagHighlighter.colors.enabledCustom`: allows users to override the current theme colors.
+- New command `goStructTagHighlighter.toggleCustomColors`: quicker way to toggle on/off the `enable custom colors` option. If you prefer to do it manually, you can still update this value in your user/workspace settings.
+
 ## v0.0.5
 - Do not crash if there is no active editor.